Two java files for login page
I haven't added this to the food management project yet as I am going to do that once I have stored all the data in the database. This is my work before the database was created. I made up a user to carry out a test to ensure it was working.